Use extmap for lazy prompt loading
Instead of trying to load the hash table of prompts mapped to valid
words wholly every time 'rom-party' was first called, instead use
extmap which manages persistence for us, and lazily loads the contents
of the map as and when required.

The initial creation of the extmap when a user first uses rom-party
still takes some time (and will do so when any of the words sources
are changed - though that's not yet implemented), but thereafter
should be instantaneous.